requirements.txt
anyio==3.6.1
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fancycompleter==0.9.1
fastapi==0.85.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
pandas==1.5.0
pathspec==0.10.1
pdbpp==0.10.3
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyrepl==0.9.0
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
wmctrl==0.4
